Extract request_token into dedicated auth utility module
